pyproject.toml
[project]
name = "toon-for-llm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cerebras-cloud-sdk>=1.59.0",
    "diskcache>=5.6.3",
    "euriai>=1.0.32",
    "google-generativeai>=0.8.5",
    "openai>=2.8.1",
    "python-dotenv>=1.2.1",
    "redis>=7.1.0",
    "sambanova>=1.2.0",
    "toon-format",
]

[tool.uv.sources]
toon-format = { git = "https://github.com/toon-format/toon-python.git" }
